refactor: share the mock-script writer between test modules
The SDK and discovery tests each carried a verbatim copy of the
ETXTBSY-safe script writer; the subtle kernel-race handling now lives
once in a test-only module.
