+++ crates/cella-container/src/discovery.rs
@@ -309,36 +309,8 @@ mod tests {
         MOCKS.get_or_init(|| {
             let dir = tempfile::TempDir::new().unwrap();
 
-            let write_script = |name: &str, body: &str| -> PathBuf {
-                use std::io::Write;
-                let path = dir.path().join(name);
-                let mut file = std::fs::File::create(&path).unwrap();
-                file.write_all(format!("#!/bin/sh\n{body}\n").as_bytes())
-                    .unwrap();
-                file.sync_all().unwrap();
-                drop(file);
-                #[cfg(unix)]
-                {
-                    use std::os::unix::fs::PermissionsExt;
-                    std::fs::set_permissions(&path, std::fs::Permissions::from_mode(0o755))
-                        .unwrap();
-                }
-                // ETXTBSY guard: the kernel may not have released the inode
-                // write reference yet (deferred __fput). Spin until exec works.
-                for _ in 0..50 {
-                    match std::process::Command::new(&path)
-                        .arg("--etxtbsy-probe")
-                        .stdout(std::process::Stdio::null())
-                        .stderr(std::process::Stdio::null())
-                        .output()
-                    {
-                        Err(e) if e.kind() == std::io::ErrorKind::ExecutableFileBusy => {
-                            std::thread::sleep(std::time::Duration::from_millis(1));
-                        }
-                        _ => break,
-                    }
-                }
-                path
+            let write_script = |name: &str, body: &str| {
+                crate::test_support::write_mock_script(dir.path(), name, body)
             };
 
             DiscMocks {
