+++ crates/cella-container/src/lib.rs
@@ -12,3 +12,5 @@ pub use backend::AppleContainerBackend;
 
 #[cfg(all(test, target_os = "macos"))]
 mod integration_tests;
+#[cfg(test)]
+mod test_support;
