+++ crates/cella-container/src/sdk/mod.rs
@@ -397,44 +397,8 @@ mod tests {
             let dir = PathBuf::from("/tmp/cella_sdk_mock_scripts");
             std::fs::create_dir_all(&dir).unwrap();
 
-            let write_script = |name: &str, body: &str| -> PathBuf {
-                use std::io::Write;
-                let path = dir.join(name);
-                let content = format!("#!/bin/sh\n{body}\n");
-                // Only write if missing or content changed; avoids ETXTBSY
-                // when another thread is executing the same file.
-                let needs_write =
-                    std::fs::read_to_string(&path).map_or(true, |existing| existing != content);
-                if needs_write {
-                    let mut file = std::fs::File::create(&path).unwrap();
-                    file.write_all(content.as_bytes()).unwrap();
-                    file.sync_all().unwrap();
-                    drop(file);
-                }
-                #[cfg(unix)]
-                {
-                    use std::os::unix::fs::PermissionsExt;
-                    let _ = std::fs::set_permissions(&path, std::fs::Permissions::from_mode(0o755));
-                }
-                // ETXTBSY guard: the kernel may not have released the inode
-                // write reference yet (deferred __fput). Spin until exec works.
-                if needs_write {
-                    for _ in 0..50 {
-                        match std::process::Command::new(&path)
-                            .arg("--etxtbsy-probe")
-                            .stdout(std::process::Stdio::null())
-                            .stderr(std::process::Stdio::null())
-                            .output()
-                        {
-                            Err(e) if e.kind() == std::io::ErrorKind::ExecutableFileBusy => {
-                                std::thread::sleep(std::time::Duration::from_millis(1));
-                            }
-                            _ => break,
-                        }
-                    }
-                }
-                path
-            };
+            let write_script =
+                |name: &str, body: &str| crate::test_support::write_mock_script(&dir, name, body);
 
             MockScripts {
                 fail: write_script("fail.sh", "exit 1"),
